+++ ListableUI/Sources/Layout/Table/TableListLayout.swift
@@ -108,7 +108,10 @@ public struct TableAppearance : ListLayoutAppearance
     
     /// How the layout should flow, either horizontally or vertically.
     public var direction: LayoutDirection
-    
+
+    /// How the list header should be positioned when content is scrolled.
+    public var listHeaderPosition: ListHeaderPosition
+
     /// If sticky section headers should be leveraged in the layout.
     public var stickySectionHeaders : Bool
     
@@ -154,6 +157,7 @@ public struct TableAppearance : ListLayoutAppearance
     /// Creates a new `TableAppearance` object.
     public init(
         direction : LayoutDirection = .vertical,
+        listHeaderPosition: ListHeaderPosition = .inline,
         stickySectionHeaders : Bool = true,
         pagingBehavior : ListPagingBehavior = .none,
         itemPositionGroupingHeight : CGFloat = 0.0,
@@ -163,6 +167,7 @@ public struct TableAppearance : ListLayoutAppearance
         layout : Layout = .init()
     ) {
         self.direction = direction
+        self.listHeaderPosition = listHeaderPosition
         self.stickySectionHeaders = stickySectionHeaders
         self.pagingBehavior = pagingBehavior
         self.itemPositionGroupingHeight = itemPositionGroupingHeight