+++ CHANGELOG.md
@@ -8,11 +8,23 @@
 
 ### Changed
 
+- `stickyListHeader` has been replaced with `listHeaderPosition`, which has three possible values: `inline`, `sticky`, and `fixed`.
+
 - `List` measurement now has options to include the `safeAreaInsets` from the `Environment` if the `contentInsetAdjustmentBehavior` will affect the layout.
 
 ### Misc
 
 # Past Releases
+
+# [4.4.0] - 2022-07-18
+
+### Fixed
+
+- Fixed a bug where child accessibility views could be exposed when they should be hidden. 
+
+### Added
+
+- Added `stickyListHeader` to layout / appearance, allowing you to pin list headers to the top of the content.
 
 # [4.3.1] - 2022-07-11
 
@@ -674,7 +686,8 @@ listActions.scrolling.scrollToSection(
 Earlier releases were ad-hoc and not tracked. To see all changes, please reference [closed PRs on Github](https://github.com/kyleve/Listable/pulls?q=is%3Apr+is%3Aclosed).
 
 
-[Main]: https://github.com/kyleve/Listable/compare/4.3.1...HEAD
+[Main]: https://github.com/kyleve/Listable/compare/4.4.0...HEAD
+[4.4.0]: https://github.com/kyleve/Listable/compare/4.3.1...4.4.0
 [4.3.1]: https://github.com/kyleve/Listable/compare/4.3.0...4.3.1
 [4.3.0]: https://github.com/kyleve/Listable/compare/4.2.0...4.3.0
 [4.2.0]: https://github.com/kyleve/Listable/compare/4.1.0...4.2.0